Show intermediate frames in navigator and enable editing translation, orientation in property editor
