+++ Gui/opensim/view/src/org/opensim/view/nodes/GroundNode.java
@@ -58,7 +58,7 @@ public class GroundNode extends OneFrameNode{
          children.add(arrNodes);
          
       }
-
+      createFrameNodes(children);
       if(children.getNodesCount()==0) setChildren(Children.LEAF);      
       addDisplayOption(displayOption.Colorable);
       addDisplayOption(displayOption.Isolatable);
