+++ Gui/opensim/view/src/org/opensim/view/nodes/OneBodyNode.java
@@ -59,7 +59,7 @@ public class OneBodyNode extends OneFrameNode{
          children.add(arrNodes);
          
       }
-
+      createFrameNodes(children);
       if(children.getNodesCount()==0) setChildren(Children.LEAF);      
       addDisplayOption(displayOption.Colorable);
       addDisplayOption(displayOption.Isolatable);
